Add files via upload
Now we are working on transition from wheeled to legged mode and back. So It does transition to legged mode but stuck at going back. There is a ring at the end of the lower joint to which we connect active string. To solve the issue of not getting back we shifted this ring upper to the middle, trying to make the motion easier for motor. But tests on the ground shows it did not help. So here is the updated code.